requirements.txt
PySide6>=6.5.0
python-pptx>=0.6.21
PyMuPDF>=1.23.0
reportlab>=4.0.0
comtypes>=1.1.14
PyInstaller>=5.10.0
pillow>=10.0.0
pywin32>=306